% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{AMRL}
\alias{AMRL}
\title{Function for estimating sigmau2 following Li and Lahiri.}
\usage{
AMRL(interval, direct, x, vardir, areanumber)
}
\arguments{
\item{interval}{interval for the algorithm.}

\item{direct}{direct estimator.}

\item{x}{matrix with explanatory variables.}

\item{vardir}{direct variance.}

\item{areanumber}{number of domains.}
}
\value{
estimated sigmau2.
}
\description{
This function estimates sigmau2 using the adjusted maximum residual
likelihood.
}
\keyword{internal}
