% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.emdi.R
\name{as.matrix.estimators.emdi}
\alias{as.matrix.estimators.emdi}
\title{Transforms estimators.emdi objects into a matrix object}
\usage{
\method{as.matrix}{estimators.emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "estimators.emdi".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Transforms estimators.emdi objects into a matrix object
}
