% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark function}
\usage{
benchmark(object, benchmark, eta, type)
}
\arguments{
\item{object}{an object of type "fh".}

\item{benchmark}{a number determining the benchmark value.}

\item{eta}{a vector containing the ratios of the population size per area and 
the total population size ($N_d/N).Values must be sorted like the domains in 
the fh object.}

\item{type}{Character indicating the type of benchmarking. Types that can be chosen
(i) Raking ("\code{raking}"),
(ii) Ratio adjustment ("\code{ratio}"),
(iii) MSE adjustment ("\code{MSE_adj}")}
}
\value{
a list with following components: 
\tabular{lll}{
     \tab \code{benchmark_data} \tab data frame containing a domain indicator 
     (Domain), direct estimates (Direct),  \cr
   \tab \tab point predictions (FH), benchmarked point predictions (FHBENCH) and 
a  \cr
\tab \tab variable indicating out-of-sample domains Out (1 for out-of-sample, 0 for 
in- \cr
\tab \tab sample) .\cr
     \tab \code{benchmark_type} \tab character returning the type of the 
     benchmark method. \cr
}
}
\description{
This function benchmarks the EBLUP estimates.
}
\references{
Datta,G. S., Ghosh, M., Steorts, R. and Maples, J. (2010) Bayesian 
benchmarking with applications to small area estimation. Test, 20, 574–588.
}
