% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown.R
\name{brown.fh}
\alias{brown.fh}
\title{Brown function}
\usage{
\method{brown}{fh}(object)
}
\arguments{
\item{object}{fitted FH model.}
}
\value{
The value W of the test statistic, the degrees of freedom and the p value.
}
\description{
This function assesses the quality of the model-based estimates by comparing 
them with the direct estimates based on a goodness-of-fit test proposed by
Brown et al. (2001).
}
\references{
Brown, G., R. Chambers, P. Heady, and D. Heasman (2001). Evaluation of small 
area estimation methods: An application to unemployment estimates from the UK
LFS. Symposium 2001 - Achieving Data Quality in a Statistical Agency: A 
Methodological Perspective, Statistics Canada.
}
