% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{coef.emdi}
\alias{coef.emdi}
\alias{coefficients}
\title{Extract model coefficients of emdi objects}
\usage{
\method{coef}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For class ebp a vector containing the residuals is returned. 
For class fh a data frame containing the realized and standardized realized 
residuals. For class direct no residuals are available.
}
\description{
Method \code{coef.emdi} extracts the model coefficients from an emdi 
object.
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

coef(emdi_model)
}
}
\seealso{
\code{\link{ebp}}, \code{\link{fh}}
}
