% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combines sample and population data}
\usage{
combine_data(pop_data, pop_domains, smp_data, smp_domains, vardir)
}
\arguments{
\item{pop_data}{a data frame with population data.}

\item{pop_domains}{population domains.}

\item{smp_data}{sample data.}

\item{smp_domains}{sample domains.}

\item{vardir}{direct variance.}
}
\value{
combined data set.
}
\description{
This function combines different data sets.
}
