% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare function}
\usage{
compare(model, ...)
}
\arguments{
\item{model}{an object of type "emdi","model".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The return of \code{compare} depends on the class of its argument. The
documentation of particular methods gives detailed information about the
return of that method.
}
\description{
Function \code{compare} is a generic function used to assess the quality of 
the model-based estimates by comparing them with the direct estimates.
}
