% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot.R
\name{compare_plot.emdi}
\alias{compare_plot.emdi}
\title{Shows plots for the comparison of estimates}
\usage{
\method{compare_plot}{emdi}(model = NULL, direct = NULL,
  indicator = "all", MSE = FALSE, CV = FALSE, label = "orig",
  color = c("blue", "lightblue3"), shape = c(16, 16),
  line_type = c("solid", "solid"), gg_theme = NULL, ...)
}
\arguments{
\item{model}{an object of type "emdi","model", representing point and MSE
estimates.}

\item{direct}{an object of type "emdi","direct", representing point
and MSE estimates. If the input argument \code{model} is of type "model","ebp",
\code{direct} is required. If the input argument \code{model} is of type 
"model","fh", the \code{direct} component is already included in the input 
argument \code{model}.}

\item{indicator}{optional character vector that selects which indicators
shall be returned: (i) all calculated indicators ("all");
(ii) each indicator name: "Mean", "Quantile_10", "Quantile_25", "Median",
"Quantile_75", "Quantile_90", "Head_Count",
"Poverty_Gap", "Gini", "Quintile_Share" or the function name/s of
"custom_indicator/s"; (iii) groups of indicators: "Quantiles", "Poverty",
"Inequality" or "Custom". If two of these groups are selected, only the first
one is returned. Note, additional custom indicators can be
defined as argument for the EBP approaches (see also \code{\link{ebp}})
and do not appear in groups of indicators even though these might belong to
one of the groups. If the \code{model} argument is of type "model","fh", 
indicator can be set to "all", "Direct", FH", or "FH_Bench" (if emdi 
object is overwritten by function benchmark). Defaults to "all".}

\item{MSE}{optional logical. If \code{TRUE}, the MSE estimates of the direct 
and model-based estimates are compared via boxplots and scatter plots.}

\item{CV}{optional logical. If \code{TRUE}, the coefficient of variation 
estimates of the direct and model-based estimates are compared via boxplots 
and scatter plots.}

\item{label}{argument that enables to customize title and axis labels. There 
are three options to label the evaluation plots: (i) original labels ("orig"), 
(ii) axis labels but no title ("no_title"), (iii) neither axis 
labels nor title ("blank").}

\item{color}{a vector with two elements. The first color determines
the color of the line in the scatter plot and the color for the direct 
estimates in the remaining plots. The second color specifies the 
color for the model-based estimates in the plots.}

\item{shape}{a numeric vector with two elements. The first shape determines
the shape of the points in the scatterplot and the shape of the points for the 
direct estimates in the remaining plots. The second shape determines 
the shape for the points for the model-based estimates. The options 
are numbered from 0 to 25.}

\item{line_type}{a character vector with two elements. The first line type 
determines the type of the line in the scatter plot and the type of the line 
for the direct estimates in the remaining plots. The second type specifies 
the line type for the model-based estimates in the plots. The options are: "twodash", 
"solid", "longdash", "dotted", "dotdash", "dashed" and "blank".}

\item{gg_theme}{\code{\link[ggplot2]{theme}} list from package \pkg{ggplot2}.
For using this argument, package \pkg{ggplot2} must be loaded via 
\code{library(ggplot2)}. See also Example 2.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scatter plot and a line plot comparing direct and model-based
estimators for each selected indicator obtained by \code{\link[ggplot2]{ggplot}}.
If the input arguments MSE and CV are set to TRUE two extra plots are created, 
respectively: the MSE/CV estimates of the direct and model-based estimates are 
compared by boxplots and scatter plots.
}
\description{
Method \code{compare_plot.emdi} produces two plots for all indicators or a 
selection of indicators for objects of type "emdi". The
first plot is a scatter plot of estimates to compare and the second is a line
plot with these estimates.
}
\details{
Since all of the comparisons need a direct estimator, the plots are 
only created for in-sample domains. For the new package version (2.0.0), the 
order of the input arguments direct and model has been changed. In this 
version (1.1.6), it is still possible to use the old order because the 
arguments are swapped internally. From the next package version on it will no 
longer be possible.
}
\examples{
\dontrun{
# Examples for comparisons of direct estimates and models of type ebp

# Loading data - population and sample data
data("eusilcA_pop")
data("eusilcA_smp")
  
# Generation of two emdi objects
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + 
self_empl + unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + 
fam_allow + house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop,
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district",
threshold = function(y){0.6 * median(y)}, L = 50, MSE = TRUE,
na.rm = TRUE, cpus = 1)

emdi_direct <- direct(y = "eqIncome", smp_data = eusilcA_smp,
smp_domains = "district", weights = "weight", threshold = 11161.44,
var = TRUE, boot_type = "naive", B = 50, seed = 123, na.rm = TRUE)

# Example 1: Receive first overview
compare_plot(model = emdi_model, direct = emdi_direct)

# Example 2: Change plot theme
library(ggplot2)
compare_plot(emdi_model, emdi_direct, indicator = "Median",
gg_theme = theme(axis.line = element_line(size = 3, colour = "grey80"),
plot.background = element_rect(fill = "lightblue3"),
legend.position = "none"))

# Example for comparison of direct estimates and models of type fh

# Loading data - population and sample data
data("eusilcA_popAgg")
data("eusilcA_smpAgg")

# Combine sample and population data -------------------------------------------
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain",
                              smp_data = eusilcA_smpAgg, smp_domains = "Domain")

# Generation of the emdi object
fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean",
             combined_data = combined_data, domains = "Domain", method = "ml", 
             MSE = TRUE)
# Example 3: Receive first overview
compare_plot(fh_std) 

# Example 4: Compare also MSE and CV estimates 
compare_plot(fh_std, MSE = TRUE, CV = TRUE)                       
}
}
\seealso{
\code{\link{emdiObject}}, \code{\link{direct}}, \code{\link{ebp}}, 
\code{\link{fh}}
}
