% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot.R
\name{compare_plot_ebp}
\alias{compare_plot_ebp}
\title{Shows plots for the comparison of estimates}
\usage{
compare_plot_ebp(model, direct, indicator = "all", MSE = FALSE,
  CV = FALSE, label = "orig", color = c("blue", "lightblue3"),
  shape = c(16, 16), line_type = c("solid", "solid"),
  gg_theme = NULL)
}
\arguments{
\item{model}{an object of type "emdi","model", representing point and MSE
estimates}

\item{direct}{an object of type "emdi","direct", representing point
and MSE estimates. If the input argument \code{model} is of type "model","ebp",
\code{direct} is required. If the input argument \code{model} is of type 
"model","fh", the \code{direct} component is already included in the input 
argument \code{model}..}

\item{indicator}{optional character vector that selects which indicators
shall be returned: (i) all calculated indicators ("all");
(ii) each indicator name: "Mean", "Quantile_10", "Quantile_25", "Median",
"Quantile_75", "Quantile_90", "Head_Count",
"Poverty_Gap", "Gini", "Quintile_Share" or the function name/s of
"custom_indicator/s"; (iii) groups of indicators: "Quantiles", "Poverty",
"Inequality" or "Custom". If two of these groups are selected, only the first
one is returned. Defaults to "all". Note, additional custom indicators can be
defined as argument for model-based approaches (see also \code{\link{ebp}})
and do not appear in groups of indicators even though these might belong to
one of the groups.}
}
\value{
A scatter plot and a line plot comparing direct and model-based
estimators for each selected indicator obtained by \code{\link[ggplot2]{ggplot}}.
If the input arguments MSE and CV are set to TRUE two extra plots are created, 
respectively: the MSE/CV estimates of the direct and model-based estimates are 
compared by boxplots and scatter plots.
}
\description{
For all indicators or a selection of indicators two plots are returned. The
first plot is a scatter plot of estimates to compare and the second is a line
plot with these estimates.
}
\keyword{internal}
