% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{confint.emdi}
\alias{confint.emdi}
\title{Confidence intervals for emdi objects}
\usage{
\method{confint}{emdi}(object, parm = "all", level = 0.95, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{parm}{equivalent to indicator argument in other methods. Optional character 
vector that selects which indicators shall be returned: (i) all calculated 
indicators ("all"); (ii) each indicator name: "Mean", "Quantile_10", 
"Quantile_25", "Median", "Quantile_75", "Quantile_90", "Head_Count",
"Poverty_Gap", "Gini", "Quintile_Share" or the function name/s of
"custom_indicator/s"; (iii) groups of indicators: "Quantiles", "Poverty",
"Inequality" or "Custom". If two of these groups are selected, only the first
one is returned. Note, additional custom indicators can be
defined as argument for model-based approaches (see also \code{\link{ebp}})
and do not appear in groups of indicators even though these might belong to
one of the groups. If the \code{model} argument is of type "model","fh", 
indicator can be set to "all", "Direct", FH", or "FH_Bench" (if emdi 
object is overwritten by function benchmark). Defaults to "all".}

\item{level}{the confidence level required. Defaults to 0.95.}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For emdi objects, a data frame with the point estimator and lower and
upper bounds is returned.
}
\description{
Method \code{confint.emdi} computes confidence intervals for direct estimates, 
empirical best linear unbiased and empirical best predictors.
}
\details{
The confidence intervals are constructed using following equation 
(exemplarily for the 95\% interval): 
\ifelse{html}{\out{CI<sub>d</sub> = y&#770;<sub>d</sub> +/- 1.96 &#8730; MSE(y&#770;<sub>d</sub>)}}{\deqn{CI_{d} = \hat{ind}_{d} +/- 1.96 \sqrt{MSE(\hat{y}_{d})}}}
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

confint(emdi_model)
}
}
\seealso{
\code{\link{direct}}, \code{\link{ebp}}, \code{\link{fh}}, 
\code{\link[stats]{confint}}
}
