% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_estimation2.R
\name{direct2}
\alias{direct2}
\title{Direct estimation of disaggregated indicators}
\usage{
direct2(y, smp_data, smp_domains = NULL, weights = NULL, sort = NULL,
  pov_line = NULL, var = FALSE, bootType = "naive", B = NULL,
  seed = NULL, X_calib = NULL, totals = NULL, na.rm = FALSE)
}
\arguments{
\item{y}{a character indicating the variable that is used for estimating the 
indicators.}

\item{smp_data}{survey data containing the above variable as well as sampling
domains, and weights and sort if selected.}

\item{smp_domains}{a character string containing the name of a variable
that indicates domains in the sample data. The variable can be numeric or a
factor.}

\item{weights}{a character containing the name of a variable for 
the sampling weights in the sample data. This argument is optional and defaults
to \code{NULL}.}

\item{sort}{a character containing the name of a variable that should be used
to sort the data. Defaults to \code{NULL}.}

\item{pov_line}{a number defining a poverty line. A poverty line is
needed for calculation e.g. of head count ratios and poverty gaps. The 
argument defaults to \code{NULL}. In this case the poverty line is set to 60\% 
of the median of the variable that is selected as y similary to the 
At-risk-of-poverty rate used in the EU (see also \cite{Social Protection Committee 2001}). 
However, any desired poverty line can be chosen.}

\item{var}{if TRUE, estimates for the variance are calcualted using a 
naive or calibrated bootstrap. Defaults to \code{FALSE}.}

\item{bootType}{a character containing the name of the bootstrap specification. 
Either a \code{"naive"} or a \code{"calibrate"} bootstrap can be used. See 
also \code{\link[laeken]{bootVar}}. Defaults to \code{naive}.}

\item{B}{a number determining the number of bootstrap populations for the 
bootstrap variance. Defaults to \code{NULL}.}

\item{seed}{an integer to set the seed for the random number generator. Random 
number generation is used in the bootstrap approach. If no seed is set, seed
is chosen randomly.}

\item{totals}{a numeric vector providing the population totals if the calibrated 
bootstrap is chosen.. Defaults to \code{NULL}. In this case, the sampling 
weights are used to calculate the totals.}

\item{na.rm}{if TRUE, observations with \code{NA} values are deleted from the 
sample data. Defaults to \code{FALSE}.}

\item{X}{a numeric matrix including calibration variables if the calibrated 
bootstrap is chosen. Defaults to NULL.}
}
\value{
An object of class "emdi" that provides direct estimators for regional
disaggregated indicators and optionally corresponding variance estimates. Generic
functions such as \code{\link{estimators}}, \code{\link{print}}, 
\code{\link{plot}}, and \code{\link{summary}} have methods that can be used
to obtain further information. See \code{\link{emdiObject}} for descriptions
of components of objects of class "emdi".
}
\description{
Function \code{direct2} estimates indicators only based on sample information. 
The variance is estimated via a naive or calibrated bootstrap. The estimation
is adapted from the estimation of direct indicators in package 
laeken.
}
\details{
The set of predefined indicators includes the mean, median, four further quantiles
(10\%, 25\%, 75\% and 90\%), head count ratio, poverty gap, Gini coefficient
and the quintile share ratio.
}
\examples{
# Loading sample data
data("eusilcA_smp")

# Example without weights and naive bootstrap
emdi_direct <- direct(y="eqIncome", smp_data=eusilcA_smp, smp_domains="district", 
weights=NULL, pov_line=10859.24, var=TRUE, bootType = "naive", B=50, 
seed=123, X = NULL, totals = NULL, na.rm=TRUE)
}
\references{
Alfons, A. and Templ, M. (2013). Estimation of Social Exclusion Indicators
from complex Surveys: The R package laeken. Journal of Statistical Software, 
54(15), 1-25.  \cr \cr
Social Protection Committee (2001). Report on indicators in the field of
poverty and social exclusions, Technical Report, European Union.
}
\seealso{
\code{\link{emdiObject}}, \code{\link[nlme]{lme}},
\code{\link{estimators.emdi}}, \code{\link{print.emdi}}, \code{\link{plot.emdi}},
\code{\link{summary.emdi}}
}

