% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1.add1.fh.R
\name{drop1.fh}
\alias{drop1.fh}
\title{drop1 method for fh function}
\usage{
\method{drop1}{fh}(object, criteria, scope, ...)
}
\arguments{
\item{object}{interval for the algorithm.}

\item{criteria}{a character string describing the model selection criterion. 
Criteria that can be chosen are "\code{AIC}", "\code{AICc}", "\code{AICb1}", 
"\code{AICb2}", "\code{BIC}", "\code{KIC}", "\code{KICc}", "\code{KICb1}", 
or "\code{KICb2}". Defaults to "\code{AIC}".}

\item{scope}{formula or a list including two formulas(\code{upper} and 
\code{lower}) specifying the models considered in the step function.}

\item{...}{further arguments passed to or from other methods.}

\item{areanumber}{number of domains.}
}
\value{
value of chosen criteria for the different variable combinations when 
one variable is dropped.
}
\description{
drop1 method for fh function
}
\keyword{internal}
