% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.emdi.R
\name{estimators.emdi}
\alias{estimators.emdi}
\title{Presents point, MSE and/or CV estimates of an emdiObject}
\usage{
\method{estimators}{emdi}(object, indicator = "all", MSE = FALSE,
  CV = FALSE, ...)
}
\arguments{
\item{object}{an object of type "emdi", representing point and,
if chosen, MSE estimates.}

\item{indicator}{optional character vector that selects which indicators
shall be returned: (i) all calculated indicators ("all");
(ii) each indicator name: "Mean" "Quantile_10", "Quantile_25", "Median",
"Quantile_75", "Quantile_90", "Head_Count", "Poverty_Gap", "Gini", 
"Quintile_Share" or the function name/s of "custom_indicator/s"; 
(ii) groups of indicators: "Quantiles", "Poverty", "Inequality" or "Custom". 
Defaults to "all". Note, additional custom indicators can be 
defined as argument for model-based approaches (\code{link{ebp}}) and do not 
appear in groups of indicators even though these might belong to one of the 
groups.}

\item{MSE}{optional logical. If TRUE, MSE estimates for selected indicators
per domain are added to the data frame of point estimates. Defaults to FALSE.}

\item{CV}{optional logical. If TRUE, coefficients of variation for selected
indicators per domain are added to the data frame of point estimates.
Defaults to FALSE.}

\item{...}{other parameters that can be passed to function \code{estimators}.}
}
\value{
an object of type "estimators.emdi" with point and/or MSE
estimates and/or calculated CV's per domain obtained from
\code{emdiObject$ind} and, if chosen, \code{emdiObject$MSE}. These objects 
contain two elements, one data frame \code{ind} and a character naming the 
indicator or indicator group \code{ind_name}.
}
\description{
Method \code{estimators.emdi} presents point and MSE estimates for regional
disaggregated indicators. Coefficients of variation are calculated
using these estimators. This method enables to select for which indicators
the estimates shall be returned. The returned object is suitable for
printing with the \code{print.estimators.emdi} method.
}
\examples{
# Loading data - population and sample data
data("eusilcA_pop")
data("eusilcA_smp")

# generate emdi object with additional indicators; here via function ebp()
set.seed(100); emdi_model <- ebp( fixed = eqIncome ~ gender + eqsize + cash + 
self_empl + unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + 
fam_allow + house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop,
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district",
pov_line = 10722.66, transformation = "box.cox", L= 1, MSE = TRUE, B = 1,
custom_indicator = list( my_max = function(y, pov_line){max(y)},
my_min = function(y, pov_line){min(y)}), na.rm = TRUE, cpus = 1)

# choose Gini coefficient and MSE and CV
estimators(emdi_model, indicator = "Gini", MSE = TRUE, CV = TRUE)

# choose custom indicators without MSE and CV
estimators(emdi_model, indicator = "Custom")
}
\seealso{
\code{\link{emdiObject}}, \code{\link{ebp}}
}

