% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eusilcA_proxmat.R
\docType{data}
\name{eusilcA_proxmat}
\alias{eusilcA_proxmat}
\title{Proximity matrix for spatial area-level models}
\format{
A data set with dimensions number of areas (94) times 
number of areas (94). Values lie between \code{0} and \code{1}. The 
respective row sums amount to 1.
}
\usage{
eusilcA_proxmat
}
\description{
A data set comprising the row-standardized proximities between the 
domains of the \code{\link{eusilcA_smpAgg}} data set.
}
\keyword{datasets}
