% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_plot}
\alias{evaluate_plot}
\title{Shows plots for the comparison of direct and model-based estimates}
\usage{
evaluate_plot(direct, model, indicator = "all", color = c("blue",
  "lightblue3"))
}
\arguments{
\item{direct}{an object of type "emdi","direct", representing point and MSE
estimates.}

\item{model}{an object of type "emdi","model", representing point and MSE
estimates.}

\item{indicator}{optional character vector that selects which indicators
shall be returned: (i) all calculated indicators ("all");
(ii) each indicator name: "Mean" "Quantile_10", "Quantile_25", "Median",
"Quantile_75", "Quantile_90", "Head_Count", 
"Poverty_Gap", "Gini", "Quintile_Share" or the function name/s of 
"custom_indicator/s"; (iii) groups of indicators: "Quantiles", "Poverty", 
"Inequality" or "Custom".If two of these groups are selected, only the first
one is returned. Defaults to "all". Note, additional custom indicators can be 
defined as argument for model-based approaches (see also \code{\link{ebp}}) 
and do not appear in groups of indicators even though these might belong to 
one of the groups.}

\item{color}{a character vector with two elements. The first color determines
the color of the line in the scatter plot and the color for the direct 
estimates in the line plot. The second color specifies the color of the line
for the model-based estimates.}
}
\value{
A scatter plot and a line plot comparing direct and model-based 
estimators for each selected indicator obtained by \code{\link{ggplot}}.
}
\description{
For all indicators or a selection of indicators two plots are returned. The 
first plot is a scatter plot of the direct and model-based point estimates 
and the second is a line plots with both point estimates.
}
\examples{
\dontrun{
# Loading data - population and sample data
data("eusilcA_pop")
data("eusilcA_smp")
  
# Generation of two emdi objects
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + 
self_empl + unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + 
fam_allow + house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop,
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district",
threshold = function(y){0.6 * median(y)}, L = 50, MSE = TRUE, B = 50,
na.rm = TRUE, cpus = 1)

emdi_direct <- direct(y = "eqIncome", smp_data = eusilcA_smp,
smp_domains = "district", weights = "weight", threshold = 10989.28,
var = TRUE, boot_type = "naive", B = 50, seed = 123, na.rm = TRUE)

# Receive first overview
evaluate_plot(direct = emdi_direct, model = emdi_model)
}
}
\seealso{
\code{\link{emdiObject}}, \code{\link{direct}}, \code{\link{ebp}},
}
