% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_test}
\alias{evaluate_test}
\title{Goodness of fit diagnostic of direct and model-based estimates}
\usage{
evaluate_test(direct, model, indicator = "all")
}
\arguments{
\item{direct}{an object of type "emdi","direct", representing point and MSE
estimates.}

\item{model}{an object of type "emdi","model", representing point and MSE
estimates.}

\item{indicator}{optional character vector that selects which indicators
shall be returned: (i) all calculated indicators ("all");
(ii) each indicator name: "Mean" "Quantile_10", "Quantile_25", "Median",
"Quantile_75", "Quantile_90", "Head_Count", 
"Poverty_Gap", "Gini", "Quintile_Share" or the function name/s of 
"custom_indicator/s"; (iii) groups of indicators: "Quantiles", "Poverty", 
"Inequality" or "Custom".If two of these groups are selected, only the first
one is returned. Defaults to "all". Note, additional custom indicators can be 
defined as argument for model-based approaches (see also \code{\link{ebp}}) 
and do not appear in groups of indicators even though these might belong to 
one of the groups.}
}
\value{
an object of type "evaluate.emdi" containing the results of the 
goodness of fit test.
}
\description{
For all indicators or a selection of indicators a goodness of fit test 
following \cite{Brown et al. (2001)} is conducted. The returned object 
is suitable for printing  with the \code{print.evaluate.emdi} method.
}
\examples{
\dontrun{
# Loading data - population and sample data
data("eusilcA_pop")
data("eusilcA_smp")
  
# Generation of two emdi objects
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + 
self_empl + unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + 
fam_allow + house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop,
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district",
threshold = function(y){0.6 * median(y)}, L = 50, MSE = TRUE, B = 50,
na.rm = TRUE, cpus = 1)

emdi_direct <- direct(y = "eqIncome", smp_data = eusilcA_smp,
smp_domains = "district", weights = "weight", threshold = 10989.28,
var = TRUE, boot_type = "naive", B = 50, seed = 123, na.rm = TRUE)

# Receive first overview
evaluate_test(direct = emdi_direct, model = emdi_model)
}
}
\references{
Brown, G., Chambers, R., Heady, P. and Heasmann, D. (2001). Evaluation of 
Small Area Estimation Methods - An Application to Unemployment Estimates from 
the UK LFS. In Proceedings of Statistics Canada Symposium 2001: Achieving Data 
Quality in a Statistical Agency: A Methodological Perspective. Statistics 
Canada. \cr \cr
}
\seealso{
\code{\link{emdiObject}}, \code{\link{direct}}, \code{\link{ebp}},
}
