% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{extractAIC.emdi}
\alias{extractAIC.emdi}
\title{Extract the AIC from a model fit of an emdi object}
\usage{
\method{extractAIC}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For objects of class fh a single number is returned. For class direct 
and ebp no AIC value is available.
}
\description{
Method \code{extractAIC.emdi} extracts the Akaike Information Criterion from 
a model fit of an emdi object.
}
\examples{
\donttest{
# Example for class fh
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain", 
                              smp_data = eusilcA_smpAgg, smp_domains = "Domain")

fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean", 
combined_data = combined_data, domains = "Domain", method = "ml", 
MSE = TRUE)

extractAIC(fh_std)
}
}
\seealso{
\code{\link{fh}}, \code{\link[stats]{extractAIC}}
}
