% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{family.emdi}
\alias{family.emdi}
\title{Extracts family object of emdi object}
\usage{
\method{family}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For objects of class ebp and fh among others characters of the family 
and link names are returned. For class direct no model family is available.
}
\description{
Method \code{family.emdi} extracts family objects from emdi objects.
}
\examples{
\donttest{
# Example for class fh
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain", 
                              smp_data = eusilcA_smpAgg, smp_domains = "Domain")

fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean", 
combined_data = combined_data, domains = "Domain", method = "ml", 
MSE = TRUE)

family(fh_std)
}
}
\seealso{
\code{\link{ebp}}, \code{\link{fh}}, \code{\link[stats]{extractAIC}}
}
