% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_robust.R
\name{fh_robust}
\alias{fh_robust}
\title{Function for Fay-Herriot model}
\usage{
fh_robust(fixed, vardir, combined_data, domains = NULL,
  correlation = "no", corMatrix = NULL, time = NULL,
  predType = "reblup", c = 1, MSE = "pseudo", B = NULL)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators.}

\item{vardir}{a character string indicating the name of the variable containing
the domain-specific sampling variances of the direct estimators that are
included in \cr \code{combined_data}.}

\item{combined_data}{a data set containing the direct estimates,
the sampling variances, the explanatory variables and the domains.}

\item{domains}{a character string indicating the domain variable that is
included in \cr \code{combined_data}. If \code{NULL}, the domains are numbered
consecutively.}

\item{correlation}{a character string determining the correlation structure.}

\item{corMatrix}{proximity matrix.}

\item{time}{a character string indicating the name of the variable containing
the time variable that is included in \code{combined_data}.}

\item{predType}{(character) one or more in c("\code{linear}", "\code{reblup}",
"\code{reblupbc}")}

\item{c}{(numeric) scalar; a multiplyer constant used in the bias correction. 
Default is to make no correction for realisations of direct estimator within
c = 1 times the standard deviation of direct estimator.}

\item{MSE}{(character) the type of the MSE. Available are "\code{pseudo}"
and "\code{robustboot}".}

\item{B}{numeric value that determines the number of bootstrap iterations.}
}
\value{
fitted robust FH model.
}
\description{
This function conducts the estimation of a Fay-Herriot model.
}
