% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_tf.R
\name{fh_tf}
\alias{fh_tf}
\title{Two-fold Fay-Herriot approach for estimating indicators at disaggregated regional levels}
\usage{
fh_tf(
  fixed,
  vardir,
  combined_data,
  domains,
  subdomains,
  subdomsize,
  transformation = "no",
  maxit = 50,
  eff_smpsize = NA,
  MSE = FALSE,
  B = 50,
  seed = 123
)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the linear mixed regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators.}

\item{vardir}{a character string indicating the name of the variable
containing the domain-specific sampling variances of the direct estimates
that are included in \cr \code{combined_data}.}

\item{combined_data}{a data set containing all the input variables that are
needed for the estimation of the two-fold Fay-Herriot model: the direct estimates,
the sampling variances, the explanatory variables, the domains, the subdomains,
and the population size. In addition, the effective sample size needs to be
included, if the arcsin transformation is chosen.}

\item{domains}{a character string indicating the domain variable that is
included in \code{combined_data}.}

\item{subdomains}{a character string indicating the subdomain variable that is
included in \code{combined_data}.}

\item{subdomsize}{a character string indicating the variable which includes
the population size at subdomain level in \code{combined_data}.}

\item{transformation}{a character that determines the type of transformation
of the dependent variable and of the sampling variances. Methods that can be
chosen (i) no transformation ("\code{no}"), (ii) log transformation
("\code{log}") of the dependent variable and of the sampling variances,
(iii) arcsin transformation ("\code{arcsin}") of the dependent variable and
of the sampling variances following. Defaults to "\code{no}".}

\item{maxit}{a number determining the maximum number of iterations for the
variance components estimation of the random effects. Defaults to 50.}

\item{eff_smpsize}{a character string indicating the name of the variable
containing the effective sample sizes that are included in
\code{combined_data}. Required argument when the arcsin transformation is
chosen. Defaults to \code{NULL}.}

\item{MSE}{if \code{TRUE}, MSE estimates are calculated based on parametric
bootstrap method. Defaults to \code{FALSE}.}

\item{B}{a single number which defines the number of bootstrap iterations
for the MSE estimation. Defaults to \code{50}.}

\item{seed}{an integer to set the seed for the random number generator. For
the usage of random number generation see details. Defaults to \code{123}.}
}
\value{
An object of class "fh_tf", "emdi" that provides estimators
for regional indicators at the domain and subdomain levels like means and
ratios and optionally MSE estimates.
}
\description{
Function \code{fh_tf} estimates indicators at two different levels
(domain and subdomain) using the two-fold Fay-Herriot approach by
\cite{Torabi and Rao (2014)}. The two-fold Fay-Herriot model is an extension
of the standard Fay-Herriot model by \cite{Fay and Herriot (1979)}. The
two-fold model is defined at the subdomain level with two random effects:
at domain level and at subdomain level. Based on the model empirical best
linear unbiased predictors (EBLUPs) and mean squared error (MSE) estimates
are provided not only at the domain level but also at the subdomain level.
The MSE estimates are obtained by the parametric bootstrap method based on
\cite{González-Manteiga et al. (2008)}. Additionally, the two-fold
Fay-Herriot approach is extended with transformations based on
\cite{Lee et al. (2024)}. The function allows to apply two different
transformations to the dependent variable: (i) logarithmic transformation
and (ii) arcsin square root transformation. For unbiased estimation of
indicators at the original scale, a bias corrected backtransformation
based on \cite{Sugasawa and Kubokawa (2017)}, \cite{Hadam et al. (2023)}, and
\cite{Lee et al. (2024)} is implemented.
}
\examples{
\donttest{
# Loading the data
data("eusilcA_Agg_mun")

# Example 1: Twofold Fay-Herriot model without transformation
fit_fh_tf <- fh_tf(fixed = Dir_Mean ~ eqsize + cash + self_empl + unempl_ben,
                   vardir = "Var_Mean", combined_data = eusilcA_Agg_mun,
                   domains = "ID_district", subdomains = "ID_municipality",
                   subdomsize = "N_ik", transformation = "no", MSE = T)

# Example 2: Twofold Fay-Herriot model with log-transformation
fit_fh_tf_log <- fh_tf(fixed = Dir_Mean ~ eqsize + cash + self_empl +
                               unempl_ben,
                       vardir = "Var_Mean", combined_data = eusilcA_Agg_mun,
                       domains = "ID_district", subdomains = "ID_municipality",
                       subdomsize = "N_ik", transformation = "log", MSE = T)

# Example 3: Twofold Fay-Herriot model with arcsine transformation
fit_fh_tf_arc <- fh_tf(fixed = Dir_HCR ~ eqsize + cash + self_empl +
                               unempl_ben,
                       vardir = "Var_HCR", combined_data = eusilcA_Agg_mun,
                       domains = "ID_district", subdomains = "ID_municipality",
                       subdomsize = "N_ik", transformation = "arcsin",
                       eff_smpsize = "eff_n", MSE = T)
}
}
\references{
Fay, R. E. and Herriot, R. A. (1979), Estimates of income for small places:
An application of James-Stein procedures to census data, Journal of the
American Statistical Association 74(366), 269-277. \cr \cr
González-Manteiga, W., Lombardía, M. J., Molina, I., Morales, D. and
Santamaría, L. (2008) Analytic and bootstrap approximations of prediction
errors under a multivariate Fay-Herriot model. Computational Statistics &
Data Analysis, 52, 5242–5252. \cr \cr
Hadam, S., Wuerz, N., Kreutzmann, A.-K., and Schmid, T. (2023), Estimating
regional unemployment with mobile network data for Functional Urban Areas in
Germany. Statistical Methods & Applications, 33, 205-233. \cr \cr
Lee, Y., Schmid, T. and Würz, N. (2024), Small area estimation using
geospatial data based on transformed two-fold nested error regression models.
Unpublished manuscript. \cr \cr
Sugasawa, S and Kubokawa, T. (2017) Transforming response values in small
area prediction. Computational Statistics & Data Analysis, 114, 47-60. \cr \cr
Torabi, M. and Rao J. N. K. (2014), On small area estimation under a sub-area
level model. Journal of Multivariate Analysis, 127, 36-55. \cr \cr
}
