% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhObject.R
\name{fhObject}
\alias{fhObject}
\title{Fitted fhObject}
\value{
The following components are always included in a fh object:
\item{\code{call}}{a list containing an image of the function call that
                   produced the object.}
\item{\code{fixed}}{a formula of fixed effects used in the linear mixed
             regression (see also \code{fixed} in \code{\link{fh}}).}
 \item{\code{framework}}{a list with following components:}
 \tabular{ll}{
     \code{direct} \tab an arranged vector containing the direct 
     estimates. \cr
     \code{vardir} \tab an arranged vector containing the direct 
     variances. \cr
     \code{N_dom_smp} \tab number of domains in the sample. \cr
     \code{N_dom_unobs} \tab number of out-of-sample domains. \cr
}
\item{\code{ind}}{data frame containing a domain indicator \code{Domain}, 
direct estimates \code{Direct}, a variable indicating out-of-sample domains 
\code{Out} (\code{1} for out-of-sample, \code{0} for in-sample) and point 
predictions (EBLUPs).}
\item{\code{method}}{a list with following components:}
\tabular{rll}{
     \tab \code{method} \tab character returning the method for estimation of
   the model parameters. \cr
     \tab \code{MSE_method} \tab character returning the method for estimation of
   the MSEs. \cr
}
\item{\code{model}}{a list with following components:}
\tabular{rll}{
     \tab \code{coefficients} \tab data frame containing the estimated regression 
coefficients, the standard errors and the \code{t}- and \code{p}-values of 
the explanatory variables. \cr
     \tab \code{variance} \tab estimated variance of the random effects. If 
correlation is selected to be "\code{spatial}" in function call, \code{variance} 
is a list containing the estimated variance of the random effects and the estimated 
spatial correlation parameter.  \cr
     \tab \code{random_effects} \tab a matrix containing the random effects 
     per domain. \cr
     \tab \code{real_residuals} \tab a matrix containing the real residuals 
     per domain. \cr
     \tab \code{std_real_residuals} \tab a matrix containing the standardised 
     real residuals per domain. \cr
     \tab \code{gamma} \tab data frame containing the shrinkage factors per 
     domain. Please note, that the shrinkage factors cannot be computed for 
     the spatial and robust extensions of the Fay-Herriot model. In those cases 
     the \code{gamma} is \code{NULL}. \cr
     \tab \code{model_select} \tab data frame containing different model 
     selection and accuracy criteria: loglikelihood, Akaike information 
     criterion (AIC), Bayesian information criterion (BIC), R2 and adjusted 
     R2 following \cite{Lahiri and Suntornchost (2015)}. Please note, that 
     the criteria cannot be computed for the robust extensions of the 
     Fay-Herriot model and the Ybarra-Lohr model. In those cases 
     \code{model_select} is \code{NULL}. \cr
     \tab \code{correlation} \tab a character naming the
                    selected correlation structure of the random effects.\cr
}
\item{\code{MSE}}{data frame containing a domain indicator \code{Domain}, 
direct variances \code{Direct}, a variable indicating out-of-sample domains 
\code{Out} (\code{1} for out-of-sample, \code{0} for in-sample) and MSE 
estimates corresponding to the point predictions in \code{ind} per domain if 
MSE is selected to be \code{TRUE} in function call. If \code{FALSE}, 
\code{MSE} is \code{NULL}.}
\item{\code{transformation}}{a list with following components:}
\tabular{rll}{
     \tab \code{transformation} \tab character returning the selected transformation
type (see also \code{transformation} in \code{\link{fh}}). \cr
     \tab \code{backtransformation} \tab character returning the type of 
     backtransformation. If \code{transformation = "no"}, 
     \code{backtransformation} is \code{NULL}. \cr
}
\item{\code{transform_param}}{\code{NULL} at the moment.}
\item{\code{successful_bootstraps}}{a matrix with domains as rows and indicators
     as columns. The cells contain the number of successful bootstraps for each 
     combination. If another MSE type than bootstrap is chosen, 
     \code{successful_bootstraps} is \code{NULL}.}
}
\description{
An object of class fh that provides estimators for regional disaggregated 
indicators like means and ratios and optionally corresponding MSE estimates. 
Generic functions such as \code{\link{estimators}}, \code{\link{print}}, 
\code{\link{plot}} and \code{\link{summary}} have methods that can be used to 
obtain further information. Additionally, for the standard Fay-Herriot model 
that is estimated via ML variance estimation a model selection function is 
provided (\code{\link{step_fh}}).
}
\references{
Lahiri, P. and Suntornchost, J. (2015), Variable selection for linear mixed
 models with applications in small area estimation, The Indian Journal of 
 Statistics 77-B(2), 312-320.
}
\seealso{
\code{\link{fh}}
}
