% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{fitted.emdi}
\alias{fitted.emdi}
\alias{fitted.values}
\title{Extract fitted values of emdi objects}
\usage{
\method{fitted}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For classes ebp and fh a vector containing the fitted values that are 
only based on the fixed effects is returned. For class direct no model fitted 
values are available.
}
\description{
Method \code{fitted.emdi} extracts the model fitted values from an emdi 
object.
}
\examples{
\donttest{
# Example for class fh
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain", 
                              smp_data = eusilcA_smpAgg, smp_domains = "Domain")
                              
fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean", 
          combined_data = combined_data, domains = "Domain", method = "ml", 
          MSE = TRUE)

fitted(fh_std)
}
}
\seealso{
\code{\link{ebp}}, \code{\link{fh}}, \code{\link[stats]{fitted}}
}
