% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{formula.emdi}
\alias{formula.emdi}
\title{Extract the model formula of an emdi object}
\usage{
\method{formula}{emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
Two-sided linear formula object describing the fixed-effects part of 
the regression model with the dependent variable on the left of a ~ operator 
and the explanatory variables on the right, separated by + operators. Method 
is not defined for emdi objects of class direct.
}
\description{
Method \code{formula.emdi} extracts the model formula of an emdi object.
}
\examples{
\donttest{
# Example for class fh
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain", 
                              smp_data = eusilcA_smpAgg, smp_domains = "Domain")
                              
fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean", 
          combined_data = combined_data, domains = "Domain", method = "ml", 
          MSE = TRUE)

formula(fh_std)
}
}
\seealso{
\code{\link{ebp}}, \code{\link{fh}}, \code{\link[stats]{formula}}
}
