% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_emdi.R
\name{getData.emdi}
\alias{getData.emdi}
\title{Extract emdi Object Data}
\usage{
\method{getData}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
Data frame used to fit the model. For classes direct and ebp the 
sample data is returned. For class fh the combined data set is returned.
}
\description{
Extracts the data frame used to fit the model.
}
\examples{
\donttest{
# Example for class direct
emdi_direct <- direct(y = "eqIncome", smp_data = eusilcA_smp, 
smp_domains = "district", weights = "weight", threshold = 11064.82, var = TRUE, 
boot_type = "naive", B = 50, seed = 123, X_calib = NULL, totals = NULL, 
na.rm = TRUE)

getData(emdi_direct)
}
}
\seealso{
\code{\link{direct}}, \code{\link{ebp}},
\code{\link{fh}}, \code{\link[nlme]{getData}}
}
