% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getGroups.emdi}
\alias{getGroups.emdi}
\title{Extract grouping factors from an emdi object}
\usage{
\method{getGroups}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A vector containing the grouping factors.
}
\description{
Method \code{getGroups.emdi} extracts grouping factors from an emdi object.
}
\examples{
\donttest{
# Example for class direct
emdi_direct <- direct(y = "eqIncome", smp_data = eusilcA_smp, 
smp_domains = "district", weights = "weight", threshold = 11064.82, var = TRUE, 
boot_type = "naive", B = 50, seed = 123, X_calib = NULL, totals = NULL, 
na.rm = TRUE)

getGroups(emdi_direct)
}
}
\seealso{
\code{\link{direct}}, \code{\link{ebp}},
\code{\link{fh}}, \code{\link[nlme]{getGroups}}
}
