% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_emdi.R
\name{getResponse.emdi}
\alias{getResponse.emdi}
\title{Extract response variable from an emdi object}
\usage{
\method{getResponse}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
Vector containing the response variable.
}
\description{
Method \code{getResponse.emdi} extracts the response variable from an emdi 
object.
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

getResponse(emdi_model)
}
}
\seealso{
\code{\link{direct}}, \code{\link{ebp}}, \code{\link{fh}}, \code{\link[nlme]{getResponse}}
}
