% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getVarCov.ebp}
\alias{getVarCov.ebp}
\title{Extract variance-covariance matrix from a fitted model of class ebp
Method \code{getVarCov.ebp} extracts the variance-covariance matrix from a fitted 
model of class ebp.}
\usage{
\method{getVarCov}{ebp}(obj, individuals, type = "random.effects", ...)
}
\arguments{
\item{obj}{an object of type "ebp".}

\item{individuals}{vector of levels of the in-sample domains can be specified 
for the types "\code{conditional}" or "\code{marginal}".}

\item{type}{a character that determines the type of variance-covariance matrix. 
Types that can be chosen
(i) random-effects variance-covariance matrix ("\code{random.effects}"),
(ii) conditional variance-covariance matrix ("\code{conditional}"), 
(iii) marginal variance-covariance matrix ("\code{marginal}"). Defaults to 
"\code{random.effects}".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A variance-covariance matrix or a list of variance-covariance matrices.
}
\description{
Extract variance-covariance matrix from a fitted model of class ebp
Method \code{getVarCov.ebp} extracts the variance-covariance matrix from a fitted 
model of class ebp.
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

getVarCov(emdi_model)
}
}
\seealso{
\code{\link{ebp}}, \code{\link[nlme]{getVarCov}}
}
