% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{getVarCov.fh}
\alias{getVarCov.fh}
\title{Extract variance-covariance matrix from a fitted model of class fh}
\usage{
\method{getVarCov}{fh}(obj, individuals, type = "random.effects", ...)
}
\arguments{
\item{obj}{an object of type "fh".}

\item{individuals}{vector of levels of the in-sample domains can be specified 
for the types "\code{conditional}" or "\code{marginal}".}

\item{type}{a character that determines the type of variance-covariance matrix. 
Types that can be chosen
(i) random-effects variance-covariance matrix ("\code{random.effects}"),
(ii) conditional variance-covariance matrix ("\code{conditional}"), 
(iii) marginal variance-covariance matrix ("\code{marginal}"). Defaults to 
"\code{random.effects}".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
A variance-covariance matrix or a list of variance-covariance matrices.
}
\description{
Method \code{getVarCov.fh} extracts the variance-covariance matrix from a fitted 
model of class fh.
}
\examples{
\donttest{
# Example for class fh
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain",
                             smp_data = eusilcA_smpAgg, smp_domains = "Domain")

fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean",
             combined_data = combined_data, domains = "Domain", method = "ml", 
             MSE = TRUE)

getVarCov(fh_std)
}
}
\seealso{
\code{\link{fh}}, \code{\link[nlme]{getVarCov}}
}
