% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shape.R
\name{load_shapeaustria_mun}
\alias{load_shapeaustria_mun}
\title{Loading the Shape File for Austrian Municipalities}
\usage{
load_shapeaustria_mun()
}
\value{
A shape file of class \code{"sf", "data.frame"}.
}
\description{
The function simplifies to load the shape file for Austrian municipalities.
}
\details{
The shape file contains the borders of Austrian municipalities.
Thus, it can be used for the visualization of estimation results for Austrian
municipalities.
}
