% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shape.R
\name{load_shapeaustria_states}
\alias{load_shapeaustria_states}
\title{Loading the Shape File for Austrian States}
\usage{
load_shapeaustria_states()
}
\value{
A shape file of class \code{"sf", "data.frame"}.
}
\description{
The function simplifies to load the shape file for Austrian states.
}
\details{
The shape file contains the borders of Austrian states.
Thus, it can be used for the visualization of estimation results for Austrian
states.
}
