% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{MPL}
\alias{MPL}
\title{Function for estimating sigmau2 using maximum likelihood.}
\usage{
MPL(interval, direct, x, vardir, areanumber)
}
\arguments{
\item{interval}{interval for the algorithm.}

\item{direct}{direct estimator.}

\item{x}{matrix with explanatory variables.}

\item{vardir}{direct variance.}

\item{areanumber}{number of domains.}
}
\value{
estimated sigmau2.
}
\description{
This function estimates sigmau2 using the maximum profile
likelihood.
}
\keyword{internal}
