% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown.R
\name{print.brown.fh}
\alias{print.brown.fh}
\title{Prints brown.fh objects}
\usage{
\method{print}{brown.fh}(x)
}
\arguments{
\item{x}{an object of type "brown.fh".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints brown.fh objects
}
