% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.compare.fh}
\alias{print.compare.fh}
\title{Prints compare.fh objects}
\usage{
\method{print}{compare.fh}(x, ...)
}
\arguments{
\item{x}{an object of type "compare.fh".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints compare.fh objects
}
