% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emdi.R
\name{print.emdi}
\alias{print.emdi}
\title{Prints an emdiObject}
\usage{
\method{print}{emdi}(x, ...)
}
\arguments{
\item{x}{an x of type "emdi", representing point and MSE
estimates following the Empirical Best Prediction by
\cite{Molina and Rao (2010)}.}

\item{...}{optional arguments passed to \code{\link{print.default}}.}
}
\description{
Basic information of an emdi object is printed.
}
\seealso{
\code{\link{emdiObject}}, \code{\link{ebp}}
}

