% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.emdi.R
\name{print.estimators.emdi}
\alias{print.estimators.emdi}
\title{Prints estimators.emdi objects}
\usage{
\method{print}{estimators.emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "estimators.emdi".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints estimators.emdi objects
}

