% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print.evaluate.emdi}
\alias{print.evaluate.emdi}
\title{Prints a evaluate.emdi object}
\usage{
\method{print}{evaluate.emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "evaluate.emdi", generally resulting
from applying evaluate_test to two objects of type "emdi", "direct" and
"emdi", "model".}

\item{...}{optional arguments passed to print.default; see the documentation on
that method functions.}
}
\description{
The results of the goodness of fit test are printed.
}
\seealso{
\code{\link{evaluate.emdi}}
}
