% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{print.step}
\alias{print.step}
\title{Prints step function results}
\usage{
\method{print}{step}(x, ...)
}
\arguments{
\item{x}{an object of type "step".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The elements described in step are printed.
}
