% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_fh.R
\name{print.step_fh}
\alias{print.step_fh}
\title{Prints step function results}
\usage{
\method{print}{step_fh}(x, ...)
}
\arguments{
\item{x}{an object of type "step_fh".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The elements described in step_fh are printed.
}
