% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.emdi.R
\name{print.summary.emdi}
\alias{print.summary.emdi}
\title{Prints a summary.emdi object}
\usage{
\method{print}{summary.emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "summary.emdi", generally resulting
from applying summary to an object of type "emdi".}

\item{...}{optional arguments passed to print.default; see the documentation on
that method functions.}
}
\description{
The elements described in summary.emdi are printed.
}
\seealso{
\code{\link{summary.emdi}}
}
