% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown.R
\name{print.syndir.fh}
\alias{print.syndir.fh}
\title{Prints syndir.fh objects}
\usage{
\method{print}{syndir.fh}(x)
}
\arguments{
\item{x}{an object of type "syndir.fh".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints syndir.fh objects
}
