% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{Reml}
\alias{Reml}
\title{Function for estimating sigmau2 with REML.}
\usage{
Reml(interval, direct, x, vardir, areanumber)
}
\arguments{
\item{interval}{interval for the algorithm.}

\item{direct}{direct estimator.}

\item{x}{matrix with explanatory variables.}

\item{vardir}{direct variance.}

\item{areanumber}{number of domains.}
}
\value{
estimated sigmau2.
}
\description{
This function estimates sigmau2.
}
\keyword{internal}
