% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{sigma.emdi}
\alias{sigma.emdi}
\title{Extract residual standard deviation of emdi objects}
\usage{
\method{sigma}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For class ebp, the estimated within-group error standard deviation is 
returned. For class fh, sigma is not available, since it is not defined for 
area-level models. For class direct, the method is not applicable.
}
\description{
Method \code{sigma.emdi} extracts the residual standard deviation from an emdi 
object.
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

sigma(emdi_model)
}
}
\seealso{
\code{\link{ebp}}
}
