% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{SML}
\alias{SML}
\title{Function for estimating sigmau2 and rho for the spatial FH model (ML).}
\usage{
SML(direct, X, vardir, areanumber, W, maxit, tol)
}
\arguments{
\item{direct}{direct estimator.}

\item{vardir}{direct variance.}

\item{areanumber}{number of domains.}

\item{maxit}{maximum number of iterations.}

\item{tol}{tolerance value for the convergence of weights.}

\item{x}{matrix with explanatory variables.}

\item{Ci}{mean squared error of x.}

\item{p}{number of covariates.}
}
\value{
estimated sigmau2 and estimated beta coefficients.
}
\description{
Function for estimating sigmau2 and rho for the spatial FH model (ML).
}
\keyword{internal}
