% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcor.R
\name{spatialcor.tests}
\alias{spatialcor.tests}
\title{Spatial autocorrelation tests}
\usage{
spatialcor.tests(direct, corMatrix)
}
\arguments{
\item{direct}{a vector containing the direct estimates of the Fay-Herriot model.}

\item{corMatrix}{proximity matrix or data frame.}
}
\value{
The values of the test statistics their corresponding p values.
}
\description{
This function computes two spatial autocorrelation tests: Moran's I and 
Geary's C.
}
