% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{SREML}
\alias{SREML}
\title{Function for estimating sigmau2 and rho for the spatial FH model (REML).}
\usage{
SREML(direct, X, vardir, areanumber, W, maxit, tol)
}
\arguments{
\item{direct}{direct estimator.}

\item{vardir}{direct variance.}

\item{areanumber}{number of domains.}

\item{W}{proximity matrix.}

\item{maxit}{maximum number of iterations for the estimation of the variance.}

\item{tol}{tolerance value for the estimation of the variance.}

\item{x}{matrix with explanatory variables.}
}
\value{
estimated sigmau2 and beta coefficients and convergence (TRUE/FALSE).
}
\description{
Function for estimating sigmau2 and rho for the spatial FH model (REML).
}
\keyword{internal}
