% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.fh.R
\name{step.fh}
\alias{step.fh}
\title{Step function}
\usage{
step.fh(object, scope, criteria = "AIC", direction = c("both",
  "backward", "forward"), trace = 1, keep = NULL, steps = 1000, ...)
}
\arguments{
\item{object}{fitted FH model.}

\item{scope}{defines the range of models examined in the stepwise search. 
This should be either a single formula, or a list containing components
\code{upper} and \code{lower}, both formulae.}

\item{direction}{the mode of stepwise search, can be one of "\code{both}",
"\code{backward}", or "\code{forward}", with a default of "\code{both}".
If the scope argument is missing the default for direction is "\code{backward}".}

\item{trace}{if positive, information is printed during the running of \code{step}.
Defaults to 1.}

\item{keep}{a filter function whose input is a fitted model object and the 
associated \code{AIC} statistic, and whose output is arbitrary. Typically 
\code{keep} will select a subset of the components of the object and return 
them. The default is not to keep anything.}

\item{steps}{the maximum number of steps to be considered. The default is 1000.}

\item{critera}{a character string describing the model selection criterion. 
Criteria that can be chosen are "\code{AIC}" or "\code{BIC}". Defaults to
"\code{AIC}".}
}
\value{
Call and coefficients of the best model due to information criterion.
}
\description{
This function selects a model by different criteria in a stepwise algorithm.
}
