% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.emdi.R
\name{subset.estimators.emdi}
\alias{subset.estimators.emdi}
\title{Subsets an estimators.emdi object}
\usage{
\method{subset}{estimators.emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "estimators.emdi".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Subsets an estimators.emdi object
}
