% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.direct.R
\name{summary.direct}
\alias{summary.direct}
\title{Summarizes an emdi direct object}
\usage{
\method{summary}{direct}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi", representing point and MSE
estimates.}

\item{...}{additional arguments that are not used in this method.}
}
\value{
an object of type "summary.direct" with information about the 
sample and population data, the usage of transformation, normality 
tests and information of the model fit.
}
\description{
Additional information about the data and model in small area estimation
methods and components of an emdi object are extracted. The returned object
is suitable for printing  with the \code{print.summary.direct} method.
}
\examples{
\donttest{
# Loading sample data
data("eusilcA_smp")

# Generation of the emdi object
emdi_direct <- direct(y = "eqIncome", smp_data = eusilcA_smp, 
smp_domains = "district", weights = "weight", threshold = 11064.82, var = TRUE, 
boot_type = "naive", B = 50, seed = 123, X_calib = NULL, totals = NULL, 
na.rm = TRUE)

# Example: Receive first overview
summary(emdi_direct)

}
}
