% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fh.R
\name{summary.fh}
\alias{summary.fh}
\title{Summarizes an emdi fh object}
\usage{
\method{summary}{fh}(object, ...)
}
\arguments{
\item{object}{an object of type "fh", representing point and MSE
estimates.}

\item{...}{additional arguments that are not used in this method.}
}
\value{
an object of type "summary.fh" with information about the 
sample and population data, the usage of transformation, normality 
tests and information of the model fit.
}
\description{
Additional information about the data and model in small area estimation
methods and components of an emdi object are extracted. The returned object
is suitable for printing  with the \code{print.summary.fh} method.
}
\examples{
\donttest{
# Example for models of type fh

# Loading data - population and sample data
data("eusilcA_popAgg")
data("eusilcA_smpAgg")

# Combine sample and population data
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain",
                              smp_data = eusilcA_smpAgg, smp_domains = "Domain")

# Generation of the emdi object
fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean",
             combined_data = combined_data, domains = "Domain", method = "ml", 
             MSE = TRUE)
             
# Example: Receive first overview
summary(fh_std)
}
}
\references{
Lahiri, P. and Suntornchost, J. (2015), Variable selection for linear mixed
models with applications in small area estimation, The Indian Journal of 
Statistics 77-B(2), 312-320. \cr \cr
Marhuenda, Y., Morales, D. and Pardo, M.C. (2014). Information criteria for 
Fay-Herriot model selection. Computational Statistics and Data Analysis 70, 
268-280. \cr \cr
Nakagawa S, Schielzeth H (2013). A general and simple method for obtaining R2 
from generalized linear mixed-effects models. Methods in Ecology and Evolution, 
4(2), 133-142.
}
\seealso{
\code{\link{emdiObject}}, \code{\link{direct}}, \code{\link{ebp}},
\code{\link{fh}}, \code{\link[MuMIn]{r.squaredGLMM}}, \code{\link[moments]{skewness}},
\code{\link[moments]{kurtosis}}, \code{\link[stats]{shapiro.test}}
}
