% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown.R
\name{syndir.fh}
\alias{syndir.fh}
\title{Synthetic vs direct}
\usage{
\method{syndir}{fh}(object)
}
\arguments{
\item{object}{fitted FH model.}
}
\value{
Correlation coefficient.
}
\description{
This function computes the correlation between the regression-synthetic part
of the Fay-Herriot model and the direct estimates.
}
