% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.emdi.R
\name{tail.estimators.emdi}
\alias{tail.estimators.emdi}
\title{Returns the last part of predicted indicators}
\usage{
\method{tail}{estimators.emdi}(x, n = 6L, addrownums = NULL, ...)
}
\arguments{
\item{x}{an object of type "estimators.emdi", representing 
point estimators and, if chosen, MSE and/or CV estimates for selected 
indicators.}

\item{n}{a single integer. If positive, it determines the number of rows for 
the data frame. If negative, all but the n last/first rows of
elements of the object.}

\item{addrownums}{if there are no row names, create them from the row numbers.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Selected rows of the object of type "estimators.emdi".
}
\description{
Returns the last part of predicted indicators
}

