% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_methods.R
\name{terms.emdi}
\alias{terms.emdi}
\title{Constructs a terms object from an emdi object}
\usage{
\method{terms}{emdi}(x, ...)
}
\arguments{
\item{x}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For classes ebp and fh a \code{\link[stats]{terms.object}} is returned. 
For class direct no terms object is available.
}
\description{
Method \code{terms.emdi} constructs a terms object from an emdi object.
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

terms(emdi_model)
}
}
\seealso{
\code{\link{ebp}}, \code{\link{fh}}, \code{\link[stats]{terms}}, 
\code{\link[stats]{terms.object}}
}
