% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{ybarralohr}
\alias{ybarralohr}
\title{Function for estimating sigmau2 and beta following Ybarra and Lohr.}
\usage{
ybarralohr(direct, x, vardir, Ci, areanumber, p, tol, maxit)
}
\arguments{
\item{direct}{direct estimator.}

\item{x}{matrix with explanatory variables.}

\item{vardir}{direct variance.}

\item{Ci}{mean squared error of x.}

\item{areanumber}{number of domains.}

\item{p}{number of covariates.}

\item{tol}{tolerance value for the convergence of weights.}

\item{maxit}{maximum number of iterations.}
}
\value{
estimated sigmau2 and estimated beta coefficients.
}
\description{
Function for estimating sigmau2 and beta following Ybarra and Lohr.
}
\keyword{internal}
