% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark function}
\usage{
benchmark(object, benchmark, eta, type, overwrite = FALSE)
}
\arguments{
\item{object}{an object of type "fh".}

\item{benchmark}{a number determining the benchmark value.}

\item{eta}{a vector containing the ratios of the population size per area and 
the total population size ($N_d/N).Values must be sorted like the domains in 
the fh object.}

\item{type}{Character indicating the type of benchmarking. Types that can be chosen
(i) Raking ("\code{raking}"),
(ii) Ratio adjustment ("\code{ratio}"),
(iii) MSE adjustment ("\code{MSE_adj}")}

\item{overwrite}{if \code{TRUE}, the EBLUP results of the fh object are 
substituted by the respective benchmarked results. Defaults
to \code{FALSE}.}
}
\value{
a data frame containing a domain indicator (Domain), direct estimates 
(Direct), point predictions (FH), benchmarked point predictions (FHBENCH) and 
a variable indicating out-of-sample domains Out (1 for out-of-sample, 0 for 
in-sample) . If overwrite is set to TRUE, the fh object is returned, but the 
point predictions of the ind data frame are substituted by the benchmarked 
results.
}
\description{
This function benchmarks the EBLUP estimates.
}
\examples{
# Loading data - population and sample data
data("eusilcA_popAgg")
data("eusilcA_smpAgg")

# Combine sample and population data -------------------------------------------
combined_data <- combine_data(pop_data = eusilcA_popAgg, pop_domains = "Domain",
                             smp_data = eusilcA_smpAgg, smp_domains = "Domain")

# Estimate Fay-Herriot model
fh_std <- fh(fixed = Mean ~ cash + self_empl, vardir = "Var_Mean",
combined_data = combined_data, domains = "Domain", method = "ml", 
interval = c(0, 100000000), MSE = TRUE)  

# Benchmark the point estimates

# Example 1: receive data frame with point estimates and their benchmarked results
fh_bench <- benchmark(fh_std, benchmark = 20140.09, 
eta = eusilcA_popAgg$ratio_n, type = "ratio")

# Example 2: overwrite the point estimates of the fh object with their benchmarked 
# results
fh_bench <- benchmark(fh_std, benchmark = 20140.09, 
eta = eusilcA_popAgg$ratio_n, type = "ratio", overwrite = TRUE)
}
\references{
Datta,G. S., Ghosh, M., Steorts, R. and Maples, J. (2010) Bayesian 
benchmarking with applications to small area estimation. Test, 20, 574–588.
}
