% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combines sample and population data}
\usage{
combine_data(pop_data, pop_domains, smp_data, smp_domains)
}
\arguments{
\item{pop_data}{a data frame with population data.}

\item{pop_domains}{a character string indicating the domain variable that is
included in \code{pop_data}.}

\item{smp_data}{a data frame with sample data.}

\item{smp_domains}{a character string indicating the domain variable that is 
included in \code{smp_data}.}
}
\value{
a combined data set.
}
\description{
This function combines the aggregated population information with the 
aggregated sample data. The merge is based on the domains. Out-of-sample 
domains will have NA values for the variables from the sample data.
}
