% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare function}
\usage{
compare(model, ...)
}
\arguments{
\item{model}{an object of type "emdi","model".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The null hypothesis, the value W of the test statistic, the degrees 
of freedom and the p value of the Brown test and the correlation coefficient 
of the synthetic part and the direct estimator.
}
\description{
Function \code{compare} is a generic function used to assess the quality of 
the model-based estimates by comparing them with the direct estimates based 
on a goodness-of-fit test proposed by Brown et al. (2001) and by computing 
the correlation between the regression-synthetic part of the model and the 
direct estimates.
}
\references{
Brown, G., R. Chambers, P. Heady, and D. Heasman (2001). Evaluation of small 
area estimation methods: An application to unemployment estimates from the UK
LFS. Symposium 2001 - Achieving Data Quality in a Statistical Agency: A 
Methodological Perspective, Statistics Canada.
}
