% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.emdi.R
\name{estimators}
\alias{estimators}
\title{Presents point, MSE and CV estimates}
\usage{
estimators(object, ...)
}
\arguments{
\item{object}{an object for which point and/or MSE estimates and/or
calculated CV's are desired.}

\item{...}{arguments to be passed to or from other methods, e.g. indicator.}
}
\value{
The return of \code{estimators} depends on the class of its argument. The
documentation of particular methods gives detailed information about the
return of that method.
}
\description{
Function \code{estimators} is a generic function used to present point and
mean squared error (MSE) estimates and calculated coefficients of variation
(CV).
}
