% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH.R
\name{fh}
\alias{fh}
\title{Standard Fay-Herriot model for disaggregated indicators}
\usage{
fh(fixed, vardir, combined_data, domains, method = "reml",
  interval = c(0, 1000), k = 10000, c = 1, transformation = "no",
  backtransformation = NULL, eff_smpsize = NULL, correlation = "no",
  corMatrix = NULL, time = NULL, Ci = NULL, tol = NULL,
  maxit = NULL, MSE = FALSE, mse_type = "analytical", B = NULL,
  alpha = 0.05)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators.}

\item{vardir}{a character string indicating the name of the variable containing
the domain-specific sampling variances of the direct estimators that are
included in \cr \code{combined_data}.}

\item{combined_data}{a data set containing the direct estimates,
the sampling variances, the explanatory variables and the domains.}

\item{domains}{a character string indicating the domain variable that is
included in \cr \code{combined_data}. If \code{NULL}, the domains are numbered
consecutively.}

\item{method}{a character string describing the method for the estimation of
the variance of the random effects. Methods that can be chosen
(i) restricted maximum likelihood (REML) method ("\code{reml}"),
(ii) maximum likelihood method ("\code{ml}"),
(iii) adjusted REML following \cite{Li and Lahiri (2010)} ("\code{amrl}"),
(iv) adjusted ML following \cite{Li and Lahiri (2010)} ("\code{ampl}"),
(v) adjusted REML following \cite{Yoshimori and Lahiri (2014)}
("\code{amrl_yl}"),
(vi) adjusted ML following \cite{Yoshimori and Lahiri (2014)}
("\code{ampl_yl}").
 Defaults to "\code{reml}".}

\item{interval}{interval for the estimation of sigmau2.}

\item{transformation}{a character that determines the type of transformation
and back-transformation. Methods that can be chosen
(i) no transformation ("\code{no}")
(ii) log transformation with naive back-transformation, i.e. simply taking the
exponential ("\code{log_naive}"),
(iii) log transformation with crude back-transformation ("\code{log_crude}"),
(iv) log transformation with Slud-Maiti back-transformation ("\code{log_SM}")
and (v) arcsin transformation with naive back-transformation ("\code{arcsin}")}

\item{backtransformation}{a character that determines the type of bracktransformation}

\item{eff_smpsize}{Effective sample size.}

\item{MSE}{if \code{TRUE}, MSE estimates are calculated. Defaults
to \code{FALSE}.}

\item{mse_type}{a character string determining the estimation method of the MSE.
Methods that can be chosen
(i) analytical MSE depending on the estimation method of the variance of the
random effect ("\code{analytical}"),
(ii) a jackknife MSE ("\code{jackknife}"),
(ii) a weighted jackknife MSE ("\code{weighted_jackknife}"),
(iii) and a bootstrap ("\code{boot}"). The latter three options are of interest
when the arcsin transformation is selected.}

\item{B}{numeric value that determines the number of bootstrap iterations.}

\item{alpha}{a numeric value that determines the confidence level for the
confidence intervals.}
}
\value{
fitted FH model.
}
\description{
Function \code{fh} estimates indicators using the Fay-Herriot approach by
\cite{Fay and Herriot (1979)}. Point estimates of indicators are
empirical best linear unbiased predictors (EBLUPs). Additionally, mean squared
error (MSE) estimation can be conducted which depends on the chosen estimation
approach for the variance of the random effect. For this estimation, six different
approaches are provided (see also \code{method}). Three different transformation
types for the dependent variable can be chosen.
}
