% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_tf.R
\name{fh_tf}
\alias{fh_tf}
\title{Two-fold Fay-Herriot Models for Indicators at a disaggregated regional level}
\usage{
fh_tf(
  fixed,
  vardir,
  combined_data,
  domains,
  subdomains,
  subdomsize,
  transformation = "no",
  maxit = 50,
  eff_smpsize = NA,
  MSE = FALSE,
  B = 50,
  seed = 123
)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the linear mixed regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators.}

\item{vardir}{a character string indicating the name of the variable
containing the domain-specific sampling variances of the direct estimates
that are included in \cr \code{combined_data}.}

\item{combined_data}{a data set containing all the input variables that are
needed for the estimation of the two-fold Fay-Herriot model: the direct estimates,
the sampling variances, the explanatory variables, the domains, the subdomains,
and the population size. In addition, the effective sample size needs to be
included, if the arcsin transformation is chosen.}

\item{domains}{a character string indicating the domain variable that is
included in \code{combined_data}.}

\item{subdomains}{a character string indicating the subdomain variable that is
included in \code{combined_data}.}

\item{subdomsize}{a character string indicating the variable which includes
the population size at subdomain level in \code{combined_data}.}

\item{transformation}{a character that determines the type of transformation
of the dependent variable and of the sampling variances. Methods that can be
chosen (i) no transformation ("\code{no}"), (ii) log transformation
("\code{log}") of the dependent variable and of the sampling variances,
(iii) arcsin transformation ("\code{arcsin}") of the dependent variable and
of the sampling variances following. Defaults to "\code{no}".}

\item{maxit}{a number determining the maximum number of iterations for the
variance components estimation of the random effects. Defaults to 50.}

\item{eff_smpsize}{a character string indicating the name of the variable
containing the effective sample sizes that are included in
\code{combined_data}. Required argument when the arcsin transformation is
chosen. Defaults to \code{NULL}.}

\item{MSE}{if \code{TRUE}, MSE estimates are calculated based on parametric
bootstrap method. Defaults to \code{FALSE}.}

\item{B}{a single number which defines the number of bootstrap iterations
for the MSE estimation.}

\item{seed}{an integer to set the seed for the random number generator. For
the usage of random number generation see details. Defaults to \code{123}.}
}
\value{
An object of class "fh_tf", "emdi" that provides estimators
for regional indicators at a disaggregated level like means and ratios and optionally
corresponding MSE estimates.
}
\description{
Two-fold Fay-Herriot Models for Indicators at a disaggregated regional level
}
