% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_methods.R
\name{fixef.emdi}
\alias{fixef.emdi}
\alias{fixed.effects}
\title{Extract fixed effects from an emdi object}
\usage{
\method{fixef}{emdi}(object, ...)
}
\arguments{
\item{object}{an object of type "emdi".}

\item{...}{additional arguments that are not used in this method.}
}
\value{
For class ebp and fh a vector containing the fixed effects is 
returned. For class direct no fixed effects are available.
}
\description{
Method \code{fixef.emdi} extracts the fixed effects from an emdi 
object.
}
\examples{
\donttest{
# Example for class ebp
emdi_model <- ebp(fixed = eqIncome ~ gender + eqsize + cash + self_empl + 
unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow + 
house_allow + cap_inv + tax_adj, pop_data = eusilcA_pop, 
pop_domains = "district", smp_data = eusilcA_smp, smp_domains = "district", 
na.rm = TRUE)

fixef(emdi_model)
}
}
\seealso{
\code{\link{ebp}}, \code{\link{fh}}
}
