% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shape.R
\name{load_shapeaustria}
\alias{load_shapeaustria}
\title{Loading the shape file for austrian district}
\usage{
load_shapeaustria()
}
\value{
A shape file of class \code{SpatialPolygonsDataFrame}.
}
\description{
Loading the shape file for austrian district
}
\details{
The shape file contains the borders of Austrian districts. Thus, it
can be used for the visualiuzation of estimation results for Austrian
districts.
}

