% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_fh.R
\name{print.step_fh}
\alias{print.step_fh}
\title{Prints step function result}
\usage{
\method{print}{step_fh}(x, ...)
}
\arguments{
\item{x}{Return from step function.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints step function result
}
