% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcor.R
\name{spatialcor.tests}
\alias{spatialcor.tests}
\title{Spatial autocorrelation tests}
\usage{
spatialcor.tests(direct, corMatrix)
}
\arguments{
\item{direct}{a vector containing direct estimates.}

\item{corMatrix}{proximity matrix or data frame.}
}
\value{
The values of the test statistics and their corresponding p values.
}
\description{
This function computes two spatial autocorrelation tests: Moran's I and 
Geary's C.
}
\details{
When creating the proximity matrix \code{corMatrix}, please make 
sure that the elements of \code{direct} and \code{corMatrix} are sorted 
equally and that \code{direct} and \code{corMatrix} do not contain any 
\code{NA}s. For a description of how to create the proximity matrix see the 
package vignette.
}
\references{
Bivand, R. (2019), spdep: Spatial Dependence: Weighting Schemes, 
Statistics. R package.
}
